Edit to FractalPlant.py
The nesting inside the grammar was not done correctly and resulted in a lot of redrawing & otherwise not wanted behaviour. This was fixed by getting the turtle's position before we recurse, and setting to this position when coming back out. Beforehand, the position was set at beginning of function, not necessarily where we start recursing